Adds parsing feature for key files.
Adds parsing feature for generated and stored key_slot_files.  Uses the
grammar defined in the flex and bison files to parse key files.  Changes
the autotools setup files to accommodate flex and bison files.  Adds a
print_keys command, which may not make it in the final release, to test
the parsing.

The parser will be eventually needed in order to verify a MAC or HMAC on
a system that does not have a Hashlet (i.e. a remote server).